Move code from main to their respective androidMain commonMain places
* jdk target 8 issue is occurring so this is a progress commit
