+++ samples/compose-samples/src/androidMain/AndroidManifest.xml
