+++ samples/compose-samples/src/androidMain/kotlin/com/squareup/sample/compose/hellocompose/AppAndroid.kt
@@ -1,0 +1,10 @@
+package com.squareup.sample.compose.hellocompose
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview(showBackground = true)
+@Composable
+private fun AppPreview() {
+  App()
+}
