+++ samples/compose-samples/src/androidMain/res/layout/legacy_view.xml
