+++ samples/compose-samples/src/androidMain/res/values/dimens.xml
