+++ samples/compose-samples/src/androidMain/res/values/strings.xml
