+++ samples/compose-samples/src/commonMain/kotlin/com/squareup/sample/compose/Defaults.kt
@@ -1,0 +1,10 @@
+package com.squareup.sample.compose
+
+import com.squareup.workflow1.RuntimeConfig
+import com.squareup.workflow1.ui.ViewEnvironment
+import com.squareup.workflow1.ui.WorkflowUiExperimentalApi
+
+expect fun defaultRuntimeConfig(): RuntimeConfig
+
+@OptIn(WorkflowUiExperimentalApi::class)
+expect fun defaultViewEnvironment(): ViewEnvironment
